+++ consensus/clique/build.gradle
@@ -47,7 +47,7 @@ dependencies {
   testImplementation project(path: ':ethereum:core', configuration: 'testSupportArtifacts')
   testImplementation project(path: ':consensus:common', configuration: 'testArtifacts')
   testImplementation project(':testutil')
-  testImplementation project(':metrics')
+  testImplementation project(':metrics:core')
 
   testImplementation 'junit:junit'
   testImplementation 'org.assertj:assertj-core'
