+++ consensus/ibft/build.gradle
@@ -47,9 +47,9 @@ dependencies {
   testImplementation project(path: ':config:', configuration: 'testSupportArtifacts')
   testImplementation project(path: ':consensus:common', configuration: 'testArtifacts')
   testImplementation project(':testutil')
-  testImplementation project(':metrics')
+  testImplementation project(':metrics:core')
 
-  integrationTestImplementation project(':metrics')
+  integrationTestImplementation project(':metrics:core')
   integrationTestImplementation 'junit:junit'
   integrationTestImplementation 'org.assertj:assertj-core'
   integrationTestImplementation 'org.mockito:mockito-core'
