+++ ethereum/blockcreation/build.gradle
@@ -27,7 +27,7 @@ dependencies {
   testImplementation project(path: ':ethereum:core', configuration: 'testSupportArtifacts')
   testImplementation project(path: ':ethereum:core', configuration: 'testArtifacts')
   testImplementation project(':testutil')
-  testImplementation project(':metrics')
+  testImplementation project(':metrics:core')
   testImplementation 'junit:junit'
   testImplementation 'org.assertj:assertj-core'
   testImplementation 'org.awaitility:awaitility'
