+++ ethereum/p2p/build.gradle
@@ -30,7 +30,7 @@ dependencies {
   implementation project(':ethereum:core')
   implementation project(':ethereum:permissioning')
   implementation project(':ethereum:rlp')
-  implementation project(':metrics')
+  implementation project(':metrics:core')
 
   implementation 'com.google.guava:guava'
   implementation 'io.prometheus:simpleclient'
