+++ metrics/build.gradle
@@ -11,37 +11,4 @@
  * specific language governing permissions and limitations under the License.
  */
 
-apply plugin: 'java-library'
-
-jar {
-  baseName 'pantheon-metrics'
-  manifest {
-    attributes(
-      'Specification-Title': baseName,
-      'Specification-Version': project.version,
-      'Implementation-Title': baseName,
-      'Implementation-Version': calculateVersion()
-      )
-  }
-}
-
-dependencies {
-  implementation project(':util')
-
-  implementation 'com.google.guava:guava'
-  implementation 'io.prometheus:simpleclient'
-  implementation 'io.prometheus:simpleclient_common'
-  implementation 'io.prometheus:simpleclient_hotspot'
-  implementation 'io.prometheus:simpleclient_pushgateway'
-  implementation 'io.vertx:vertx-core'
-  implementation 'io.vertx:vertx-web'
-  implementation 'org.apache.logging.log4j:log4j-api'
-
-  runtime 'org.apache.logging.log4j:log4j-core'
-
-  // test dependencies.
-  testImplementation 'junit:junit'
-  testImplementation 'org.assertj:assertj-core'
-  testImplementation 'org.mockito:mockito-core'
-  testImplementation 'com.squareup.okhttp3:okhttp'
-}
+jar { enabled = false }
