+++ metrics/core/build.gradle
@@ -14,7 +14,7 @@
 apply plugin: 'java-library'
 
 jar {
-  baseName 'services-util'
+  baseName 'pantheon-metrics-core'
   manifest {
     attributes(
       'Specification-Title': baseName,
@@ -26,15 +26,22 @@ jar {
 }
 
 dependencies {
-  api project(':util')
-  implementation project(':metrics:core')
+  implementation project(':util')
 
-  implementation 'org.apache.logging.log4j:log4j-api'
   implementation 'com.google.guava:guava'
-  implementation 'org.rocksdb:rocksdbjni'
+  implementation 'io.prometheus:simpleclient'
+  implementation 'io.prometheus:simpleclient_common'
+  implementation 'io.prometheus:simpleclient_hotspot'
+  implementation 'io.prometheus:simpleclient_pushgateway'
+  implementation 'io.vertx:vertx-core'
+  implementation 'io.vertx:vertx-web'
+  implementation 'org.apache.logging.log4j:log4j-api'
 
   runtime 'org.apache.logging.log4j:log4j-core'
 
+  // test dependencies.
   testImplementation 'junit:junit'
   testImplementation 'org.assertj:assertj-core'
+  testImplementation 'org.mockito:mockito-core'
+  testImplementation 'com.squareup.okhttp3:okhttp'
 }
