+++ metrics/core/src/main/java/tech/pegasys/pantheon/metrics/MetricCategory.java
