+++ metrics/core/src/main/java/tech/pegasys/pantheon/metrics/noop/NoOpCounter.java
