+++ metrics/core/src/main/java/tech/pegasys/pantheon/metrics/prometheus/PrometheusCounter.java
