+++ metrics/core/src/main/java/tech/pegasys/pantheon/metrics/vertx/VertxMetricsAdapterFactory.java
