+++ services/kvstore/build.gradle
@@ -28,7 +28,8 @@ jar {
 dependencies {
   api project(':util')
 
-  implementation project(':metrics')
+  implementation project(':metrics:core')
+  implementation project(':metrics:rocksdb')
   implementation project(':services:util')
 
   implementation 'com.google.guava:guava'
