+++ services/kvstore/src/main/java/tech/pegasys/pantheon/services/kvstore/RocksDbKeyValueStorage.java
@@ -16,6 +16,8 @@ import tech.pegasys.pantheon.metrics.Counter;
 import tech.pegasys.pantheon.metrics.MetricCategory;
 import tech.pegasys.pantheon.metrics.MetricsSystem;
 import tech.pegasys.pantheon.metrics.OperationTimer;
+import tech.pegasys.pantheon.metrics.prometheus.PrometheusMetricsSystem;
+import tech.pegasys.pantheon.metrics.rocksdb.RocksDBStats;
 import tech.pegasys.pantheon.services.util.RocksDbUtil;
 import tech.pegasys.pantheon.util.bytes.BytesValue;
 
@@ -75,7 +77,7 @@ public class RocksDbKeyValueStorage implements KeyValueStorage, Closeable {
                   MetricCategory.KVSTORE_ROCKSDB,
                   "read_latency_seconds",
                   "Latency for read from RocksDB.",
-                  rocksDbConfiguration.getLabel())
+                  "database")
               .labels(rocksDbConfiguration.getLabel());
       removeLatency =
           metricsSystem
@@ -83,7 +85,7 @@ public class RocksDbKeyValueStorage implements KeyValueStorage, Closeable {
                   MetricCategory.KVSTORE_ROCKSDB,
                   "remove_latency_seconds",
                   "Latency of remove requests from RocksDB.",
-                  rocksDbConfiguration.getLabel())
+                  "database")
               .labels(rocksDbConfiguration.getLabel());
       writeLatency =
           metricsSystem
@@ -91,7 +93,7 @@ public class RocksDbKeyValueStorage implements KeyValueStorage, Closeable {
                   MetricCategory.KVSTORE_ROCKSDB,
                   "write_latency_seconds",
                   "Latency for write to RocksDB.",
-                  rocksDbConfiguration.getLabel())
+                  "database")
               .labels(rocksDbConfiguration.getLabel());
       commitLatency =
           metricsSystem
@@ -99,8 +101,12 @@ public class RocksDbKeyValueStorage implements KeyValueStorage, Closeable {
                   MetricCategory.KVSTORE_ROCKSDB,
                   "commit_latency_seconds",
                   "Latency for commits to RocksDB.",
-                  rocksDbConfiguration.getLabel())
+                  "database")
               .labels(rocksDbConfiguration.getLabel());
+
+      if (metricsSystem instanceof PrometheusMetricsSystem) {
+        RocksDBStats.registerRocksDBMetrics(stats, (PrometheusMetricsSystem) metricsSystem);
+      }
 
       metricsSystem.createLongGauge(
           MetricCategory.KVSTORE_ROCKSDB,
@@ -121,7 +127,7 @@ public class RocksDbKeyValueStorage implements KeyValueStorage, Closeable {
                   MetricCategory.KVSTORE_ROCKSDB,
                   "rollback_count",
                   "Number of RocksDB transactions rolled back.",
-                  rocksDbConfiguration.getLabel())
+                  "database")
               .labels(rocksDbConfiguration.getLabel());
     } catch (final RocksDBException e) {
       throw new StorageException(e);
