+++ services/pipeline/build.gradle
@@ -27,7 +27,7 @@ jar {
 
 dependencies {
   api project(':util')
-  implementation project(':metrics')
+  implementation project(':metrics:core')
 
   implementation 'org.apache.logging.log4j:log4j-api'
   implementation 'com.google.guava:guava'
