+++ settings.gradle
@@ -14,7 +14,6 @@
 rootProject.name='pantheon'
 include 'acceptance-tests'
 include 'config'
-include 'consensus'
 include 'consensus:clique'
 include 'consensus:common'
 include 'consensus:ibft'
@@ -32,7 +31,8 @@ include 'ethereum:permissioning'
 include 'ethereum:referencetests'
 include 'ethereum:rlp'
 include 'ethereum:trie'
-include 'metrics'
+include 'metrics:core'
+include 'metrics:rocksdb'
 include 'pantheon'
 include 'services:kvstore'
 include 'services:pipeline'
